requirements.txt
google-i18n-address
google.cloud
nameparser
pandas